Align `home-dir` with original scripts
The original bash script uses the `HOME` environment variable[1] to determine the user's home
directory while we were using thee `user.home` JVM property on non-windows platforms. This turns out
to not be equivalent (see e.g. JDK-8280357[2] for a recent issue caused by this). So instead, we now
also use the environment variable.

As it turns out, the original Powershell script also first checks the `HOME` environment variable[3]
before falling back to `USERPROFILE` (which is what we were using so far), so we also align that to
ensure equivalent behavior. This also means we can drop the "workaround" comment here since it no
longer is one.

Closes #114.

[1]  https://github.com/clojure/brew-install/blob/b03428b1b31a8d60573a284d6971fbd3db7b748d/src/main/resources/clojure/install/clojure#L280
[2]  https://bugs.openjdk.org/browse/JDK-8280357
[3]  https://github.com/clojure/brew-install/blob/dabb82bd011b0510006dd88464f350052adb69b5/src/main/resources/clojure/install/ClojureTools.psm1#L254-L258